+++ Kiosk/src/Main.java
@@ -1,0 +1,8 @@
+package Kiosk;
+
+public class Main {
+	public static void main(String[] args) {
+		mainGUI test = new mainGUI();
+		test.window1();
+	}
+}
