--- Main.java
@@ -1,8 +1,0 @@
-package test;
-
-public class Main {
-	public static void main(String[] args) {
-		mainGUI test = new mainGUI();
-		test.window1();
-	}
-}
