--- 자바 수업 메모.txt
@@ -1,28 +1,0 @@
-(휴게소)키오스크 
-
-1. 메인페이지 (메인) => **메인창1** (먹고가기, 포장하기) 패널1
-: 가장 기본 화면
-1-1. 상품 리스트(카테고리) (메인) 패널2
-: 상품들을 카테고리별로 보여줌
-- 휴게소 음식 패널2
-
-1. 분식 : 떡볶이, 김밥, 라면, 쫄면, 만두(김치, 고기) => 패널3
-2. 일식 : 초밥, 우동, 돈까스, 모밀(온/냉) => 패널4
-3. 양식 : 스파게티, 피자, 햄버거 => 패널5
-4. 한식 : 비빔밥, 불고기, 갈비탕, 육개장 => 패널6
-2. 상품 선택 후 옵션 띄우기 (리스트) (서브1) => **상품 선택1, 2, ...**
-: 상품에서 고를 수 있는 옵션 보여줌
-3. 옵션 선택 (서브1-1) => **상품 옵션 1-1, 2-1...**
-: 체크 박스, 라디오 버튼
-
-4. 장바구니에 담기 (서브2)
-4-1. 추가 및 제거 (서브2-1)
-: 지금까지 담은 상품 종류 및 가격 보여줌
-5. 결제 (서브3) => **결제창**
-: 결제 방법 선택 및 결제 확인
-
-5-1. 최종 가격(서브4)
-6. 결제 완료 시 영수증, 홈 화면 복귀(장바구니 초기화) (서브5, 메인복귀)
-=> **완료창**
-
-
